requirements.txt
ujson
nujson
orjson
simplejson
hjson
python-rapidjson